Add files via upload
Competition Task • Objective: Design and implement an algorithmic trading strategy that selects only one trading target and modes (long or short) and design the most profitable trading strategy based on market conditions. • Competition setting and rules: • Each team may consist of 2, 3 or 4 participants. • Every team begins with a $10,000 budget and has the flexibility to allocate any amount, high or low, on a single stock. • The remaining funds will be invested in a fixed deposit with the annual interest rate of 4%. • The team that accumulates the highest surplus as the winner of the game..Dataset • The dataset includes stocks, ETFs, and crypto currencies. • Each trading target prices, including open price, close price, highest price, lowest price and volume, are provided.